refactor: 입찰 application layer 패키지 이동 및 네이밍 변경
