--- backend/ddang/src/main/java/com/ddang/ddang/bid/application/dto/BidDto.java
@@ -1,6 +1,0 @@
-package com.ddang.ddang.bid.application.dto;
-
-import com.ddang.ddang.auction.domain.Auction;
-
-public record BidDto(Long previousBidderId, Auction auction, String auctionImageAbsoluteUrl) {
-}
