+++ backend/ddang/src/main/java/com/ddang/ddang/bid/application/event/dto/NotificationPreviousBidDto.java
@@ -1,0 +1,6 @@
+package com.ddang.ddang.bid.application.event.dto;
+
+import com.ddang.ddang.auction.domain.Auction;
+
+public record NotificationPreviousBidDto(Long previousBidderId, Auction auction, String auctionImageAbsoluteUrl) {
+}
