+++ backend/ddang/src/main/java/com/ddang/ddang/bid/presentation/BidController.java
@@ -3,8 +3,8 @@ package com.ddang.ddang.bid.presentation;
 import com.ddang.ddang.authentication.configuration.AuthenticateUser;
 import com.ddang.ddang.authentication.domain.dto.AuthenticationUserInfo;
 import com.ddang.ddang.bid.application.BidService;
-import com.ddang.ddang.bid.application.dto.CreateBidDto;
-import com.ddang.ddang.bid.application.dto.ReadBidDto;
+import com.ddang.ddang.bid.application.dto.request.CreateBidDto;
+import com.ddang.ddang.bid.application.dto.response.ReadBidDto;
 import com.ddang.ddang.bid.presentation.dto.request.CreateBidRequest;
 import com.ddang.ddang.bid.presentation.dto.response.ReadBidResponse;
 import com.ddang.ddang.image.presentation.util.ImageRelativeUrlFinder;
