+++ backend/ddang/src/test/java/com/ddang/ddang/bid/presentation/fixture/BidControllerFixture.java
@@ -1,7 +1,7 @@
 package com.ddang.ddang.bid.presentation.fixture;
 
 import com.ddang.ddang.authentication.infrastructure.jwt.PrivateClaims;
-import com.ddang.ddang.bid.application.dto.ReadBidDto;
+import com.ddang.ddang.bid.application.dto.response.ReadBidDto;
 import com.ddang.ddang.bid.presentation.dto.request.CreateBidRequest;
 import com.ddang.ddang.configuration.CommonControllerSliceTest;
